readme: update compilation instructions
Compiling on Windows is not possible based on the current instructions.
Add additional instructions regarding prerequisites that need to be
manually installed and the necessary flags for the build tools so these
are also found during compilation. Restructure the installing from
source part to separate the two different type of instructions (i.e.
Windows from everything else with more sane development environment).