feat: Add OpenSelection action, and honor case in the configuration bindings
This patch adds the OpenSelection action, which can be used to quickly
open a link in the buffer. With an appropriate word selector, one can
select an URL and open it with a keybind.

It is a variation of this patch:
https://st.suckless.org/patches/open_copied_url/, which I use daily to
quickly open URLs in chat, mails, diffs, etc.

I have not tested it, but it should work with other URIs, like file://
and such.

I've also fixed a minor behavior I encoutered while developing this
patch: By default, the Config will upper() the keybind used when it is a
simple key, but for simple Alt+Key binds (in my case, I use Alt+o for
this feature), the event sent to the terminal is Alt+o (Alt+6F), while
the configuration stores Alt+O (Alt+4F). I've corrected this behavior,
which might break stuff; but it seemed more logical to me this way.

Signed-off-by: Louis Solofrizzo <louis@ne02ptzero.me>