Bugfix DECRQM (Dec Request Mode) response
Interactive terminal sequence response for DEC Private Mode Query is
invalid -- it may respond with a different mode and status than the one
requested.

This causes programs to fail or mismatch contour's terminal
capabilities. In my API I was expecting the requested mode value in the
response pattern, and, not discovering it after some timeout, wrongly
reported that Contour does not support any DEC Private Modes at all.

This bug is present in all versions of Contour supporting DEC Private
Mode, since 256e313d and refactored in faa66e7d, the use of:

       static_cast<DECMode>(mode)

Is invalid, because the DECMode enum has "holes". It is not contiguous.

A new 'fromDECModeNum' function is provided, returning an optional
DecMode, and, existing function 'isValidDECMode' is refactored to check
whether the optional value returned by 'fromDECModeNum' has_value().

Given the provided file `test/display-modes.sh`, we can make inquiry of
DEC Private modes and see that DEC modes 1 through 27 are returned with
non-matching response modes mismatched by enum item order and status::

    $ ./test/display-modes.sh
    Query mode 1: response=$'\E[?2;2$y' after 13ms
    Query mode 2: response=$'\E[?3;2$y' after 11ms
    Query mode 3: response=$'\E[?4;2$y' after 10ms
    Query mode 4: response=$'\E[?5;2$y' after 11ms
    Query mode 5: response=$'\E[?9;2$y' after 11ms
    Query mode 6: response=$'\E[?1000;2$y' after 12ms
    Query mode 7: response=$'\E[?1001;2$y' after 13ms
    Query mode 8: response=$'\E[?1002;2$y' after 13ms
    Query mode 9: response=$'\E[?1003;2$y' after 12ms
    Query mode 10: response=$'\E[?1048;2$y' after 14ms

After this bugfix, this is corrected:

    $ ./test/display-modes.sh
    Query mode 1: response=$'\E[?1;2$y' after 11ms
    Query mode 2: response=$'\E[?2;2$y' after 11ms
    Query mode 3: response=$'\E[?3;2$y' after 8ms
    Query mode 4: response=$'\E[?4;2$y' after 10ms
    Query mode 5: response=$'\E[?5;2$y' after 8ms
    Query mode 6: response=$'\E[?6;2$y' after 11ms
    Query mode 7: response=$'\E[?7;1$y' after 11ms
    Query mode 8: response=$'\E[?8;0$y' after 10ms
    Query mode 9: response=$'\E[?9;2$y' after 10ms
    Query mode 10: response=$'\E[?10;2$y' after 12ms

Further, looking at other uses of 'isValidDECMode', some were unnecessary,
that they have the DECMode type bound to the variable is sufficient
to describe that mode as valid for Contour, the check is superfluous:

      isValidDECMode(static_cast<unsigned int>(mode)));